+++ app/pages/f7ios/sortable-list.vue
@@ -4,6 +4,7 @@
       <div class="navbar-inner">
         <div class="left sliding"><a class="back link" href="#"><i class="icon icon-back"></i><span>Back</span></a></div>
         <div class="center sliding">Sortable List</div>
+        <div class="right"><a href="#" class="link toggle-sortable icon-only">Edit</a></div>
       </div>
     </div>
     <div class="page-content">