+++ CHANGELOG.md
@@ -1,12 +1,20 @@
 # Changelog
 
+## Next
+
+### New features
+
+### Bug-fixes
+
+### Improvements
+
 ## Version 1.4.4 (released on 2017-05-08)
 
 ### Bug-fixes
 
-- #489 - Cordova platform installation failed for Android
-- #495 - Cordova failed to open .idea/modules.xml file
+- [#489 - Cordova platform installation failed for Android](https://github.com/scriptPilot/app-framework/issues/489)
+- [#495 - Cordova failed to open .idea/modules.xml file](https://github.com/scriptPilot/app-framework/issues/495)
 
 ### Improvements
 
-- #429 - Deploy directly to device during tests
+- [#429 - Deploy directly to device during tests](https://github.com/scriptPilot/app-framework/issues/429)