+++ CHANGELOG.md
@@ -14,6 +14,7 @@
 - [#403 - UI components - Overlay login: No way to close it](https://github.com/scriptPilot/app-framework/issues/403)
 - [#406 - UI components - Scrollable tabbar does not work](https://github.com/scriptPilot/app-framework/issues/406)
 - [#496 - Console error on dynamic page load (view.history undefined)](https://github.com/scriptPilot/app-framework/issues/496)
+- [#499 - JS files not compressed in 1.4.11](https://github.com/scriptPilot/app-framework/issues/499)
 
 ### Improvements