+++ CHANGELOG.md
@@ -6,6 +6,7 @@
 
 ### Bug-fixes
 
+- [#110 - Public list has sometimes trouble with multiple access the same time](https://github.com/scriptPilot/app-framework/issues/110)
 - [#387 - Calendar timeline is not scrollable](https://github.com/scriptPilot/app-framework/issues/387)
 - [#394 - UI components - Sortable list: Edit button not visible](https://github.com/scriptPilot/app-framework/issues/394)
 - [#400 - UI components - Infinite scroll does not work](https://github.com/scriptPilot/app-framework/issues/400)